Breaks up verify.py into smaller more focused modules.
Resolves: https://jira.duraspace.org/browse/FCREPO-2451
